Many things

+++ bin/baseq_stats.py
@@ -8,45 +8,70 @@ import ggplot
 
 def main(args):
 
-    sample = args.sample
-    q_scores = list()
+    if args.fastq is not None:
+        sample = args.sample
+        q_scores = list()
 
-    for fastq in args.fastq:
-        if args.gz:
-            os.system('gunzip -c ' + fastq + " > basq_stats.tmp.fastq")
-            fastq_file = 'basq_stats.tmp.fastq'
-        else:
-            fastq_file = fastq
+        for fastq in args.fastq:
+            if args.gz:
+                os.system('gunzip -c ' + fastq + " > basq_stats.tmp.fastq")
+                fastq_file = 'basq_stats.tmp.fastq'
+            else:
+                fastq_file = fastq
 
-        seqio = SeqIO.parse(fastq_file, format='fastq')
+            seqio = SeqIO.parse(fastq_file, format='fastq')
 
-        for seq_record in seqio:
-            q_scores += seq_record.letter_annotations['phred_quality']
+            for seq_record in seqio:
+                q_scores += seq_record.letter_annotations['phred_quality']
 
-    df = pandas.DataFrame(q_scores,columns=['BaseQuality'])
+        df = pandas.DataFrame(q_scores,columns=['BaseQuality'])
 
-    p = ggplot.ggplot(ggplot.aes(x='BaseQuality'), data=df) + ggplot.geom_histogram(binwidth=1) + ggplot.labs(title = sample)
-    p.save(args.prefix + '-hist.png')
+        p = ggplot.ggplot(ggplot.aes(x='BaseQuality'), data=df) + ggplot.geom_histogram(binwidth=1) + ggplot.labs(title = sample)
+        p.save(args.prefix + '-hist.png')
 
-    # statistics
-    fout = open(args.prefix + '-statistics.csv','w')
-    for i in np.arange(0.1,1,0.1):
-        fout.write(str(int(100*i)) + '% quantile,' + str(df.quantile(i)[0]) + '\n')
-    for i in np.arange(1,41):
-        fout.write('Proportion bases >= Q' + str(i) + ',' + str((df>=i).sum()[0]/float(df.shape[0])) + '\n')
+        # statistics
+        fout = open(args.prefix + '-statistics.csv','w')
+        for i in np.arange(0.1,1,0.1):
+            fout.write(str(int(100*i)) + '% quantile,' + str(df.quantile(i)[0]) + '\n')
+        for i in np.arange(1,41):
+            fout.write('Proportion bases >= Q' + str(i) + ',' + str((df>=i).sum()[0]/float(df.shape[0])) + '\n')
 
-    fout.close()
+        fout.close()
 
-    if os.path.exists('basq_stats.tmp.fastq'):
-        os.system('rm basq_stats.tmp.fastq')
+        if os.path.exists('basq_stats.tmp.fastq'):
+            os.system('rm basq_stats.tmp.fastq')
+    else:
+        import pysam
+        sample = args.sample
+        q_scores = list()
+
+        bam = pysam.AlignmentFile(args.bam,'rb')
+
+        for seq_record in bam:
+            q_scores += pysam.qualitystring_to_array(seq_record.qual).tolist()
+
+        df = pandas.DataFrame(q_scores,columns=['BaseQuality'])
+
+        p = ggplot.ggplot(ggplot.aes(x='BaseQuality'), data=df) + ggplot.geom_histogram(binwidth=1) + ggplot.labs(title = sample)
+        p.save(args.prefix + '-hist.png')
+
+        # statistics
+        fout = open(args.prefix + '-statistics.csv','w')
+        for i in np.arange(0.1,1,0.1):
+            fout.write(str(int(100*i)) + '% quantile,' + str(df.quantile(i)[0]) + '\n')
+        for i in np.arange(1,41):
+            fout.write('Proportion bases >= Q' + str(i) + ',' + str((df>=i).sum()[0]/float(df.shape[0])) + '\n')
+
+        fout.close()
 
 
-parser = argparse.ArgumentParser(description='Compute and plot BaseQuality statistics')
+parser = argparse.ArgumentParser(description='Compute and plot BaseQuality statistics from FastQ or BAM file')
 parser.add_argument(
     '--fastq',
     type=str,
-    required=True,
+    required=False,
     nargs='+',
+    default=None,
     help='Space-delimited list of on or more Fastq files'
 )
 parser.add_argument(
@@ -54,6 +79,13 @@ parser.add_argument(
     action='store_true',
     required=False,
     help='If fastq files are compressed (.gz)'
+)
+parser.add_argument(
+    '--bam',
+    type=str,
+    required=False,
+    default=None,
+    help='BAM file'
 )
 parser.add_argument(
     '--sample',
@@ -69,5 +101,9 @@ parser.add_argument(
     help='Output prefix name'
 )
 args = parser.parse_args()
+
+if args.bam is not None and args.fastq is not None:
+    print 'Supply only BAM or FastQ file!'
+    sys.exit()
 
 main(args=args)
